change for comments because the destroy method seemed like a bad hack.
git-svn-id: http://svn.joyent.com/opensource/connector/source/trunk@383 7a7ed7af-43cf-e7d3-831c-e7cc5cc1171a

+++ app/controllers/comments_controller.rb
@@ -31,6 +31,11 @@ class CommentsController < AuthenticatedController
 
   def remove
     comment = Comment.find(params[:id], :scope => :read) # don't have polymorphic delete scope yet
+    
+    unless comment.user == current_user || comment.commentable.owner == current_user || curren_user.admin?
+      return render :nothing => true
+    end
+    
     item = comment.commentable
     comment.destroy
